requirements.txt
Flask==1.1.2
google-cloud-speech==2.0.0